requirements.txt
python-telegram-bot==21.10
yt-dlp>=2026.3.3
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
